switch to integer-based packet identifiers

+++ network/main/java/tech/fastj/network/client/ClientListener.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 import java.net.DatagramPacket;
 
 public interface ClientListener {
-    void receiveTCP(byte identifier, Client client) throws IOException;
+    void receiveTCP(int identifier, Client client) throws IOException;
 
     void receiveUDP(DatagramPacket packet, Client client) throws IOException;
 }
